% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_heatmap.R
\name{histheat}
\alias{histheat}
\title{A 1d heatmap histogram}
\usage{
histheat(x, bins = 30, x_name = "value", narrow = TRUE, x_limits = NULL, ...)
}
\arguments{
\item{x}{the vector of points to be used in the histogram.}

\item{bins}{the number of bins for the histogram.  Defaults to 30.}

\item{x_name}{the name to put on the x-axis.}

\item{narrow}{defaults to TRUE.  Makes the chart narrow on the y-axis.}

\item{x_limits}{the range of the x-axis to plot. Defaults to all with NULL.}

\item{...}{values that work in ggplot2::scale_x_continous. Mostly for the breaks argument.}
}
\description{
A ggplot2 based function to provide a simple heatmap histogram.
}
\examples{
x_vector <-  rnorm(1000)
histheat(x_vector)
histheat(x_vector, narrow = FALSE, bins = 20)
histheat(x_vector, x_limits = c(-3, 3), breaks = seq(-3,3, by = .5))
}
